lesson5/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
curlify2==2.0.0
deepdiff==8.0.1
Faker==33.0.0
fastapi==0.115.2
future==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.2
orderly-set==5.2.2
outcome==1.3.0.post0
packaging==24.2
playwright==1.48.0
pluggy==1.5.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyee==12.0.0
PySocks==1.7.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
ruff==0.7.1
selene==2.0.0rc9
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.40.0
text-unidecode==1.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
wsproto==1.2.0
